fix: Loading special URP/HDRP shaders at runtime does not fail anymore. (#116)
* fix: URP clearcoat shader loading at runtime.

The shader name contained a folder and thus was malformed.

fix: Switched from asset-path-based to GUID-based shader loading (in the Editor) in order to allow for a flexible folder layout without risking breaks/regressions should the layout change in the future.

* fix: HDRP stack-lit shader loading at runtime.

The shader name contained a folder and thus was malformed.

* fix: Switched from asset-path-based to GUID-based shader loading (in the Editor 2021 and newer) in order to allow for a flexible folder layout without risking breaks/regressions should the layout change in the future.

* fix: Removed redundant qualifier

* fix: Avoid expensive UnityEngine.Object null check when accessing cached default shaders.